fix(builder): fib example now works as expected

+++ examples/fib/fib.rs
@@ -1,4 +1,4 @@
 #[no_mangle]
-pub fn fib(a: u32, b: u32, n: u32) -> u32 {
+pub fn fib(a: u128, b: u128, n: u128) -> u128 {
     if n == 0 { a } else { fib(b, a + b, n - 1) }
 }
